Remove def pre-collection
The original intent was to be able to refer to defs from other libs,
since we didn't have access to library ids before we finished lowering
the library itself.

However, this complicates include expansion since we'd essentially have
to do it twice, instead of just doing it once during main lowering.

Now that we can refer to external libraries during lowering, we don't
need to collect defs beforehand anymore.