Add `PackageFile`
This points to a specific instance of a `SourceFile` which is part
of a given `Package`, as a `SourceFile` may be part of certain
packages.

This permits parent item lookups since they eventually need to know
what package the item is a part of in order to find the starting
place to find nodes from.