Base `AstId`s on hashed entities (#142)
This gives better stability for item locations in a file than the old toposort visit method, in that:

- Items of the same name but different AstId kind don't invalidate each other
- Adding/removing differently names items does not invalidate the rest of the AstIdMap

This is based off of r-a's AstIdMap.

This also consolidates ItemLocMap into ItemCollection (now ChildItems) as it's still useful to have a more stable representation not built directly from the syntax tree.